requirements.txt
fastapi
uvicorn
httpx
joblib
pandas
numpy
scikit-learn
python-dotenv
pydantic
